+++ lib/wed/mode-tree.ts
@@ -1,0 +1,488 @@
+/**
+ * Manage a tree of modes.
+ * @author Louis-Dominique Dubeau
+ * @license MPL 2.0
+ * @copyright Mangalam Research Center for Buddhist Languages
+ */
+import * as Ajv from "ajv";
+import * as mergeOptions from "merge-options";
+
+import { toGUISelector } from "./domutil";
+import { Mode } from "./mode";
+import { Mode as ModeOption } from "./options";
+import { Runtime } from "./runtime";
+import { ModeValidator } from "./validator";
+import { ComplexAttributesSpec, WedOptions } from "./wed-options";
+import * as wedOptionsSchema from "./wed-options-schema.json";
+
+/**
+ * A callback for reporting wed option errors.
+ *
+ * @param path The mode's path, as specified in the configuration.
+ *
+ * @param errors The errors encountered.
+ */
+type WedOptionsErrorCallback = (path: string, errors: string[]) => void;
+
+export interface CleanedWedOptions extends WedOptions {
+  attributes: ComplexAttributesSpec;
+}
+
+export interface AttributeHidingSpecs {
+  elements: {
+    selector: string,
+    attributes: (string | { except: string[]})[];
+  }[];
+}
+
+/**
+ * A node for the mode tree.
+ */
+class ModeNode {
+  private _attributeHidingSpecs: AttributeHidingSpecs | null;
+
+  /**
+   * @param mode The mode that this node holds.
+   *
+   * @param selector The selector that determines to what this modes apply. This
+   * selector must have been converted to operate in the GUI tree.
+   *
+   * @param submodes The submodes set for this mode.
+   */
+  constructor(public readonly mode: Mode<{}>,
+              public readonly selector: string,
+              public readonly submodes: ModeNode[],
+              public readonly wedOptions: CleanedWedOptions) {}
+
+  /**
+   * Determines whether an element matched by the selector of this ``ModeNode``
+   * node in the GUI tree contains a node. If it does, this means that the mode
+   * that this ``ModeNode`` holds, or one of the submode, governs the node.
+   *
+   * @param parentScope The element from which the selector in this ``ModeNode``
+   * is interpreted.
+   *
+   * @param node A GUI node to test.
+   *
+   * @returns The element that represents the top of the mode's region of
+   * activity and contains ``node``. Returns ``null`` if no element contains the
+   * node.
+   */
+  containingElement(parentScope: Element, node: Node): Element | null {
+    if (!parentScope.contains(node)) {
+      return null;
+    }
+
+    if (this.selector === "") {
+      return parentScope;
+    }
+
+    const regions = parentScope.querySelectorAll(this.selector);
+    for (const region of Array.from(regions)) {
+      if (region.contains(node)) {
+        return region;
+      }
+    }
+
+    return null;
+  }
+
+  reduceTopFirst<T>(fn: (accumulator: T, node: ModeNode) => T,
+                    initialValue: T): T {
+    let value = fn(initialValue, this);
+
+    for (const submode of this.submodes) {
+      value = fn(value, submode);
+    }
+
+    return value;
+  }
+
+  get attributeHidingSpecs(): AttributeHidingSpecs | null {
+    if (this._attributeHidingSpecs === undefined) {
+      const attributeHiding = this.wedOptions.attributes.autohide;
+      if (attributeHiding === undefined) {
+        // No attribute hiding...
+        this._attributeHidingSpecs = null;
+      }
+      else {
+        const method = attributeHiding.method;
+        if (method !== "selector") {
+          throw new Error(`unknown attribute hiding method: ${method}`);
+        }
+
+        const specs: AttributeHidingSpecs = {
+          elements: [],
+        };
+
+        for (const element of attributeHiding.elements) {
+          const copy = mergeOptions({}, element);
+          copy.selector =
+            toGUISelector(copy.selector,
+                          this.mode.getAbsoluteNamespaceMappings());
+          specs.elements.push(copy);
+        }
+
+        this._attributeHidingSpecs = specs;
+      }
+    }
+
+    return this._attributeHidingSpecs;
+  }
+}
+
+interface ModeConstructor {
+  new (editor: Editor, modeOptions: {}): Mode<{}>;
+}
+
+interface ModeModule {
+  Mode: ModeConstructor;
+}
+
+export interface Editor {
+  runtime: Runtime;
+  gui_root: Element;
+  data_root: Element;
+  fromDataNode(node: Node): Node;
+}
+
+/**
+ * A tree containing the modes configured for the current editing session.
+ */
+export class ModeTree {
+  private root: ModeNode;
+  private readonly runtime: Runtime;
+  private readonly wedOptionsValidator: Ajv.ValidateFunction =
+    new Ajv().compile(JSON.parse(wedOptionsSchema));
+  private cachedMaxLabelNode: ModeNode;
+
+  /**
+   * @param editor The editor for which we are building this tree.
+   *
+   * @param option The ``mode`` option from the options passed to the wed
+   * instance. This object will construct a tree from this option.
+   */
+  constructor(private readonly editor: Editor,
+              private readonly option: ModeOption) {
+    this.runtime = editor.runtime;
+  }
+
+  /**
+   * Load the modes, initalize them and build the tree.
+   *
+   * @returns A promise that resolves to ``this`` once all the modes are loaded
+   * and initialized.
+   */
+  async init(): Promise<ModeTree> {
+    const combinedErrors: string[] = [];
+    this.root = await this.makeNodes(
+      "",
+      this.option,
+      (path: string, errors: string[]) => {
+        for (const error of errors) {
+          combinedErrors.push(
+            `mode at path ${path} has an error in its wed options: ${error}`);
+        }
+      });
+
+    if (combinedErrors.length > 0) {
+      throw new Error(`wed options are incorrect: ${combinedErrors.join("")}`);
+    }
+
+    return this;
+  }
+
+  /**
+   * Make the nodes of the tree. This function operates recursively: it will
+   * inspect ``option`` for a ``submode`` option and will call itself to create
+   * the necessary child nodes.
+   *
+   * @param selector The selector associated with the options passed in the 2nd
+   * argument.
+   *
+   * @param option The mode option being processed.
+   *
+   * @param errorHanler The handler to call on errors in processing the wed
+   * options. If this handler is called at all, then the returned value should
+   * not be used. We operate this way because we want to report all errors that
+   * can be reported, rather than abort early.
+   *
+   * @returns A promise that resolves to the created node.
+   */
+  private async makeNodes(selector: string,
+                          option: ModeOption,
+                          errorHandler: WedOptionsErrorCallback):
+  Promise<ModeNode> {
+    const submode = option.submode;
+    const mode = await this.initMode(option.path, option.options);
+    const submodes = (submode !== undefined) ?
+      [await this.makeNodes(toGUISelector(submode.selector,
+                                          mode.getAbsoluteNamespaceMappings()),
+                            submode.mode,
+                            errorHandler)] :
+      [];
+    const rawOptions = mode.getWedOptions();
+    const result = this.processWedOptions(rawOptions);
+    let cleanedOptions: CleanedWedOptions;
+    if (Array.isArray(result)) {
+      errorHandler(option.path, result);
+      // This is a lie.
+      cleanedOptions = rawOptions as CleanedWedOptions;
+    }
+    else {
+      cleanedOptions = result;
+    }
+    return new ModeNode(mode, selector, submodes, cleanedOptions);
+  }
+
+  private async initMode(path: string,
+                         options: {} | undefined = {}): Promise<Mode<{}>> {
+    const mmodule: ModeModule = await this.loadMode(path);
+    const mode = new mmodule.Mode(this.editor, options);
+
+    await mode.init();
+    return mode;
+  }
+
+  private async loadMode(path: string): Promise<ModeModule> {
+    const runtime = this.runtime;
+    try {
+      return (await runtime.resolveModules(path))[0] as ModeModule;
+    }
+    // tslint:disable-next-line:no-empty
+    catch (ex) {}
+
+    if (path.indexOf("/") !== -1) {
+      // It is an actual path so don't try any further loading.
+      throw new Error(`can't load mode ${path}`);
+    }
+
+    path = `./modes/${path}/${path}`;
+
+    try {
+      return (await runtime.resolveModules(path))[0] as ModeModule;
+    }
+    // tslint:disable-next-line:no-empty
+    catch (ex) {}
+
+    try {
+      return (await runtime.resolveModules(`${path}-mode`))[0] as ModeModule;
+    }
+    // tslint:disable-next-line:no-empty
+    catch (ex) {}
+
+    return (await runtime.resolveModules(`${path}_mode`))[0] as ModeModule;
+  }
+
+  /**
+   * Validates and normalizes the options to a specific format.
+   *
+   * @param options The raw options obtained from the mode.
+   *
+   * @returns The cleaned options if successful. If there were error the return
+   * value is an array of error messages.
+   */
+  private processWedOptions(options: WedOptions): CleanedWedOptions | string[] {
+    const errors: string[] = [];
+
+    const ovalidator = this.wedOptionsValidator;
+    const valid = ovalidator(options);
+    if (!(valid as boolean)) {
+      if (ovalidator.errors !== undefined) {
+        for (const error of ovalidator.errors) {
+          errors.push(`${error.dataPath} ${error.message}`);
+        }
+      }
+
+      return errors;
+    }
+
+    const max = options.label_levels.max;
+
+    const initial = options.label_levels.initial;
+
+    // We cannot validate this with a schema.
+    if (initial > max) {
+      errors.push("label_levels.initial must be <= label_levels.max");
+    }
+
+    if (options.attributes === undefined) {
+      options.attributes = "hide";
+    }
+
+    // Normalize the format of options.attributes.
+    if (typeof options.attributes === "string") {
+      const tmp = options.attributes;
+      // We need the type cast at the end because otherwise TS infers a type of
+      // { handling: "hide" | "show" | "edit" }.
+      options.attributes = {
+        handling: tmp,
+      } as { handling: "hide" } | { handling: "show" | "edit" };
+    }
+
+    if (errors.length !== 0) {
+      return errors;
+    }
+
+    return options as CleanedWedOptions;
+  }
+
+  /**
+   * Get the mode that governs a node.
+   *
+   * @param The node we want to check. This must be a done in the data tree or
+   * the GUI tree.
+   *
+   * @returns The mode that governs the node.
+   */
+  getMode(node: Node): Mode<{}> {
+    return this.getModeNode(node).mode;
+  }
+
+  /**
+   * Get the processed wed options that are in effect for a given node.
+   *
+   * @param The node we want to check. This must be a done in the data tree or
+   * the GUI tree.
+   *
+   * @returns The wed options that governs the node.
+   */
+  getWedOptions(node: Node): CleanedWedOptions {
+    const modeNode = this.getModeNode(node);
+    return modeNode.wedOptions;
+  }
+
+  /**
+   * Get the attribute handling that applies to a specific node.
+   */
+  getAttributeHandling(node: Node): "show" | "hide" | "edit" {
+    return this.getWedOptions(node).attributes.handling;
+  }
+
+  /**
+   * Get the attribute hiding specs that apply to a specific node.
+   */
+  getAttributeHidingSpecs(node: Node):  AttributeHidingSpecs | null {
+    return this.getModeNode(node).attributeHidingSpecs;
+  }
+
+  /**
+   * Get the mode node that governs a node.
+   *
+   * @param The node we want to check. This must be a done in the data tree or
+   * the GUI tree.
+   *
+   * @returns The mode that governs the node.
+   */
+  private getModeNode(node: Node): ModeNode {
+    // Handle the trivial case where there is no submode first.
+    if (this.root.submodes.length === 0) {
+      return this.root;
+    }
+
+    if (this.editor.data_root.contains(node)) {
+      node = this.editor.fromDataNode(node);
+    }
+
+    if (!this.editor.gui_root.contains(node)) {
+      throw new Error("did not pass a node in the GUI or data tree");
+    }
+
+    const result = this._getModeNode(this.root, this.editor.gui_root, node);
+    if (result === undefined) {
+      throw new Error("cannot find a mode for the node; something is wrong");
+    }
+
+    return result;
+  }
+
+  private _getModeNode(parent: ModeNode, parentScope: Element,
+                       node: Node): ModeNode | undefined {
+    const scope = parent.containingElement(parentScope, node);
+    if (scope !== null) {
+      let narrower: ModeNode | undefined;
+      for (const submode of parent.submodes) {
+        narrower = this._getModeNode(submode, scope, node);
+        if (narrower !== undefined) {
+          return narrower;
+        }
+      }
+
+      return parent;
+    }
+
+    return undefined;
+  }
+
+  /**
+   * Get the stylesheets that the modes define. It is up to the mode to use
+   * stylesheets that are written so as to avoid interfering with one another.
+   *
+   * @returns The list of sheets used by the modes. Straight duplicates are
+   * eliminated from the list. The paths must not require any further
+   * interpretation from wed.
+   */
+  getStylesheets(): string[] {
+    return Object.keys(this.root.reduceTopFirst(
+      (accumulator: Record<string, boolean>, node) => {
+        for (const sheet of node.mode.getStylesheets()) {
+          accumulator[sheet] = true;
+        }
+        return accumulator;
+      }, Object.create(null)));
+  }
+
+  /**
+   * Get the maximum label visibility level configured by the modes. This
+   * function looks at all modes in use and returns the highest number it finds.
+   *
+   * @returns The maximum label visibility level.
+   */
+  getMaxLabelLevel(): number {
+    return this.maxLabelLevelNode.wedOptions.label_levels.max;
+  }
+
+  /**
+   * Get the initial label visibility level configured by the modes. This
+   * function looks at all modes in use and returns the number that is set by
+   * the same mode used to provide the value of [[getMaxLabelLevel]].
+   *
+   * @returns The initial label visibility level.
+   */
+  getInitialLabelLevel(): number {
+    return this.maxLabelLevelNode.wedOptions.label_levels.initial;
+  }
+
+  /**
+   * The node with the maximum label visibility level. If multiple nodes have
+   * the same value, the earlier node "wins", and is the one provided by this
+   * property. For instance, if the root node and its submode have the same
+   * number, then this property has the root node for value.
+   *
+   * This is a cached value, computed on first access.
+   */
+  private get maxLabelLevelNode(): ModeNode {
+    if (this.cachedMaxLabelNode === undefined) {
+      this.cachedMaxLabelNode = this.root.reduceTopFirst<ModeNode>(
+        (accumulator, node) => {
+          const accMax = accumulator.wedOptions.label_levels.max;
+          const nodeMax = node.wedOptions.label_levels.max;
+          return (nodeMax > accMax) ? node : accumulator;
+        }, this.root);
+    }
+
+    return this.cachedMaxLabelNode;
+  }
+
+  /**
+   * @returns The list of all mode validators defined by the modes.
+   */
+  getValidators(): ModeValidator[] {
+    return this.root.reduceTopFirst<ModeValidator[]>(
+      (accumulator, node) => {
+        const validator = node.mode.getValidator();
+        return validator !== undefined ?
+          accumulator.concat(validator) : accumulator;
+      }, []);
+  }
+}